% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drawVector}
\alias{drawVector}
\title{Draw Vector Function}
\usage{
drawVector(buffer)
}
\arguments{
\item{buffer}{A vector of elements to be printed}
}
\value{
No return value, prints to screen
}
\description{
This function prints all the elements of a vector
}
\examples{
\dontrun{
data <- c(1:12)
drawVector(data)
}
}
