% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getUserAction}
\alias{getUserAction}
\title{Get User Action Function}
\usage{
getUserAction()
}
\value{
A vector entered by the user
}
\description{
This function gets the buffer introduced by the user. Typically a numerical vector.
}
\examples{
\dontrun{
vector <- getUserAction()
}
}
