% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initImages}
\alias{initImages}
\title{Initialize Images Function}
\usage{
initImages(image_name)
}
\arguments{
\item{image_name}{Name of the image file in inst/images}
}
\value{
No return value
}
\description{
This function displays an image from the package resources.
}
